fix warning
git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/trunk@1052 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ examples/vehiclemmi/vehiclemmi_core.ss/src/VehicleMmiCppQt/qtworkaround.cpp
@@ -93,7 +93,7 @@ namespace VehicleMmiCppQt
                 *a = wchar_t(u);
                 ++a;
             }
-            return a - pArray;
+            return static_cast<int>(a - pArray);
         }
     }
